requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
